refactor: pulled all API calls into a single service
