+++ app/facades/movie_facade.rb
@@ -6,20 +6,20 @@ class MovieFacade
   end
   
   def self.top_cast(id)
-    cast = TopCastService.get_cast(id)[:cast][0..9]
+    cast = MovieService.get_cast(id)[:cast][0..9]
     cast.map do |data|
       Actor.new(data)
     end
   end
 
   def self.top_rated
-    TopRatedMovieService.movies[:results].map do |movie|
+    MovieService.top_rated_movies[:results].map do |movie|
       MovieDetail.new(movie)
     end
   end
 
   def self.movie_search(query)
-    MovieSearchService.search(query)[:results].map do |movie|
+    MovieService.search(query)[:results].map do |movie|
       MovieDetail.new(movie)
     end
   end
