+++ app/services/movie_service.rb
@@ -1,12 +1,23 @@
 class MovieService
   def self.get_movie(id)
-    response = connection.get("movie/#{id}")
-    JSON.parse(response.body, symbolize_names: true)
+    JSON.parse(conn.get("movie/#{id}").body, symbolize_names: true)
   end
 
+   def self.top_rated_movies
+    JSON.parse(conn.get("movie/top_rated").body, symbolize_names: true)
+  end
+
+  def self.get_cast(id)
+    JSON.parse(conn.get("movie/#{id}/credits").body, symbolize_names: true)
+  end
+
+  def self.search(query)
+    JSON.parse(conn.get("search/movie?query=#{query}").body, symbolize_names: true)
+  end
+  
   private
   
-  def self.connection
+  def self.conn
     Faraday.new( 
       url: "https://api.themoviedb.org/3/",
       params: { api_key: ENV['MOVIE_DB_KEY'] }
